+++ marimo/__init__.py
@@ -34,6 +34,8 @@ __all__ = [
     "output",
     "plain_text",
     "pdf",
+    "redirect_stderr",
+    "redirect_stdout",
     "refs",
     "right",
     "state",
@@ -64,7 +66,12 @@ from marimo._plugins.stateless.plain_text import plain_text
 from marimo._plugins.stateless.tabs import tabs
 from marimo._plugins.stateless.tree import tree
 from marimo._runtime import output
-from marimo._runtime.capture import capture_stderr, capture_stdout
+from marimo._runtime.capture import (
+    capture_stderr,
+    capture_stdout,
+    redirect_stderr,
+    redirect_stdout,
+)
 from marimo._runtime.control_flow import MarimoStopError, stop
 from marimo._runtime.runtime import defs, refs
 from marimo._runtime.state import state
